requirements.txt
anyio==3.6.1
backports.cached-property==1.0.2
click==8.1.3
colorama==0.4.5
dnspython==2.2.1
graphql-core==3.2.1
h11==0.13.0
idna==3.3
Pygments==2.13.0
pymongo==4.2.0
python-dateutil==2.8.2
python-multipart==0.0.5
six==1.16.0
sniffio==1.2.0
starlette==0.20.4
strawberry-graphql==0.126.0
typing-extensions==4.3.0
uvicorn==0.18.2
